pyproject.toml
[tool.poetry]
name = "nonebot_plugin_biliav"
version = "1.0.12"
description = "NoneBot biliav小程序 查看插件"
authors = ["knva <txt_ch@163.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9.0"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
httpx = ">=0.18.0,<1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
